test: break helpers out into the tests/Helpers dir
